Fixing typo rm old SchemaEditor
