+++ frontend/src/components/management/renderers/SchemaEntryRenderer.svelte
@@ -278,7 +278,7 @@
         color="success"
         size="sm"
         class="justify-content-center text-center py-0 px-1"
-        active={"view" == tab_option}
+        active={"view" === tab_option}
         title={$_("view")}
         on:click={() => (tab_option = "view")}
       >
@@ -289,7 +289,7 @@
         color="success"
         size="sm"
         class="justify-content-center text-center py-0 px-1"
-        active={"edit_meta" == tab_option}
+        active={"edit_meta" === tab_option}
         title={$_("edit") + " meta"}
         on:click={() => (tab_option = "edit_meta")}
       >
@@ -301,7 +301,7 @@
           color="success"
           size="sm"
           class="justify-content-center text-center py-0 px-1"
-          active={"edit_content" == tab_option}
+          active={"edit_content" === tab_option}
           title={$_("edit") + " payload"}
           on:click={() => (tab_option = "edit_content")}
         >
@@ -313,7 +313,7 @@
         color="success"
         size="sm"
         class="justify-content-center text-center py-0 px-1"
-        active={"attachments" == tab_option}
+        active={"attachments" === tab_option}
         title={$_("attachments")}
         on:click={() => (tab_option = "attachments")}
       >
@@ -324,7 +324,7 @@
         color="success"
         size="sm"
         class="justify-content-center text-center py-0 px-1"
-        active={"history" == tab_option}
+        active={"history" === tab_option}
         title={$_("history")}
         on:click={() => (tab_option = "history")}
       >
@@ -411,7 +411,7 @@
         class="px-1 pb-1 h-100"
         style="text-align: left; direction: ltr; overflow: hidden auto;"
       >
-        <SchemaEditor bind:content bind:items />
+        <SchemaEditor bind:content />
         {#if errorContent}
           <h3 class="mt-3">Error:</h3>
           <Prism bind:code={errorContent} />
